+++ packages/google_maps_flutter/google_maps_flutter_platform_interface/CHANGELOG.md
@@ -1,6 +1,6 @@
 ## 1.0.5
 
-* Add a `fromJson` constructor to `BitmapDescriptor` so serialized descriptors can be synchronously re-hydrated.
+* Temporarily add a `fromJson` constructor to `BitmapDescriptor` so serialized descriptors can be synchronously re-hydrated. This will be removed when a fix for [this issue](https://github.com/flutter/flutter/issues/70330) lands.
 
 ## 1.0.4